foi adicionado funções que dão opções de qual calculo vai querer, ligado com a outra pagina de calculo de armazenamento
